[scheduler cleanup phase 1]: Move NodeInfo utils into pkg/scheduler/cache
